Rename(#34): ColorCircle 패키지 생성
