+++ src/components/PalettePanel/PalettePanel.tsx
@@ -1,6 +1,6 @@
 import * as S from './PalettePanel.styled.ts';
 
-import ColorCircle                      from '@/components/Common/ColorCircle';
+import ColorCircle from '@/components/Common/ColorCircle/ColorCircle.ts';
 import { LabelColors, LabelColorsType } from '@/styles/colorThemes.ts';
 
 type PalettePanelProps = {
