+++ src/components/group/invited/InvitedGroupInfo.tsx
@@ -11,9 +11,9 @@ import {
 } from '@/api/group/deleteGroupInvite';
 import patchGroupInviteAccepted from '@/api/group/patchGroupInviteAccepted';
 import Check from '@/assets/icons/check.svg';
-import X           from '@/assets/icons/x.svg';
-import ColorCircle from '@/components/Common/ColorCircle';
-import Tooltip     from '@/components/Tooltip/Tooltip';
+import X from '@/assets/icons/x.svg';
+import ColorCircle from '@/components/Common/ColorCircle/ColorCircle';
+import Tooltip from '@/components/Tooltip/Tooltip';
 import { LabelColorsType } from '@/styles/colorThemes.ts';
 
 type InvitedGroupInfoProps = {
