+++ src/components/group/update/GroupUpdatePanel.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 
-import PalettePanel     from '@/components/PalettePanel/PalettePanel';
-import ColorCircle      from '@/components/Common/ColorCircle';
+import PalettePanel from '@/components/PalettePanel/PalettePanel';
+import ColorCircle from '@/components/Common/ColorCircle/ColorCircle';
 import { InputDefault } from '@/components/Common/InputText';
-import PopOver          from '@/components/Common/PopOver';
+import PopOver from '@/components/Common/PopOver';
 
 import { GroupUpdateInfo } from '@/@types/Group';
 import { SearchMemberForCreateGroupData } from '@/@types/Member';
