+++ src/components/navbar/sheet/member/MemberPalette.tsx
@@ -4,10 +4,10 @@ import {
   patchMemberLabelColor,
   UpdateLabelColorRequest,
 } from '@/api/member/patchMemberLabelColor';
-import check                            from '@/assets/icons/check.svg';
+import check from '@/assets/icons/check.svg';
 import { LabelColors, LabelColorsType } from '@/styles/colorThemes.ts';
-import ColorCircle                      from '@/components/Common/ColorCircle';
-import useMemberStore                   from '@/stores/MemberStore';
+import ColorCircle from '@/components/Common/ColorCircle/ColorCircle';
+import useMemberStore from '@/stores/MemberStore';
 
 function MemberPalette() {
   const { mainColor, setMainColor: setLabelColor } = useMemberStore();
